ai-code-assistant/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-code-assistant"
version = "0.4.0"
description = "Cognify AI - AI-powered code assistant with code review, generation, editing, semantic search, knowledge base, and usage analytics"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Ashok", email = "akkssy@example.com"}
]
keywords = ["ai", "code-review", "code-generation", "langchain", "ollama", "cli", "cognify", "semantic-search", "knowledge-base"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-ollama>=0.1.0",
    "langchain-core>=0.1.0",
    # Multi-provider LLM support
    "langchain-openai>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-groq>=0.1.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # RAG / Codebase Search
    "chromadb>=1.0.0",
    "sentence-transformers>=2.0.0",
    "watchdog>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
cognify = "ai_code_assistant.cli:main"
ai-assist = "ai_code_assistant.cli:main"

[project.urls]
Homepage = "https://github.com/akkssy/architects-handbook/tree/main/ai-code-assistant"
Repository = "https://github.com/akkssy/architects-handbook"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

